在windows自动编译平台增加android自动编译

+++ ThirdLibrary/build_script/build_libtiff.sh
@@ -40,23 +40,27 @@ CUR_DIR=`pwd`
 
 #下载源码:
 if [ ! -d ${RABBITIM_BUILD_SOURCE_CODE} ]; then
-    TIFF_VERSION=4.0.6
-    echo "wget -q ftp://ftp.remotesensing.org/pub/libtiff/tiff-${TIFF_VERSION}.zip"
-    mkdir -p ${RABBITIM_BUILD_SOURCE_CODE}
-    cd ${RABBITIM_BUILD_SOURCE_CODE}
-    wget -q ftp://ftp.remotesensing.org/pub/libtiff/tiff-${TIFF_VERSION}.zip
-    unzip tiff-${TIFF_VERSION}.zip
-    mv tiff-${TIFF_VERSION} ..
-    rm -fr *
-    cd ..
-    rm -fr ${RABBITIM_BUILD_SOURCE_CODE}
-    mv -f tiff-${TIFF_VERSION} ${RABBITIM_BUILD_SOURCE_CODE}
+    if [ "TRUE" = "${RABBITIM_USE_REPOSITORIES}" ]; then
+        git clone https://github.com/KangLin/tiff.git ${RABBITIM_BUILD_SOURCE_CODE} 
+    else
+        echo "wget -c -nv https://github.com/KangLin/tiff/archive/master.zip"
+        mkdir -p ${RABBITIM_BUILD_SOURCE_CODE}
+        cd ${RABBITIM_BUILD_SOURCE_CODE}
+        wget -c -nv https://github.com/KangLin/tiff/archive/master.zip
+        unzip -q master.zip
+        mv tiff-master ..
+        rm -fr *
+        cd ..
+        rm -fr ${RABBITIM_BUILD_SOURCE_CODE}
+        mv -f tiff-master ${RABBITIM_BUILD_SOURCE_CODE}
+    fi
 fi
 
 cd ${RABBITIM_BUILD_SOURCE_CODE}
-
-if [ "$RABBITIM_CLEAN" ]; then
-    make clean
+mkdir -p build_${RABBITIM_BUILD_TARGERT}
+cd build_${RABBITIM_BUILD_TARGERT}
+if [ "$RABBITIM_CLEAN" = "TRUE" ]; then
+    rm -fr *
 fi
 
 echo ""
@@ -72,67 +76,48 @@ echo "RABBITIM_BUILD_STATIC:$RABBITIM_BUILD_STATIC"
 echo ""
 
 echo "configure ..."
-MAKE=make
+MAKE_PARA="-- ${RABBITIM_MAKE_JOB_PARA}"
 if [ "$RABBITIM_BUILD_STATIC" = "static" ]; then
-    CONFIG_PARA="--enable-static --disable-shared --without-ld-shared"
-    export LDFLAGS="-static"
+    CMAKE_PARA="-DBUILD_SHARED_LIBS=OFF"
 else
-    CONFIG_PARA="--disable-static --enable-shared"
+    CMAKE_PARA="-DBUILD_SHARED_LIBS=ON"
 fi
 case ${RABBITIM_BUILD_TARGERT} in
     android)
-        export CC=${RABBITIM_BUILD_CROSS_PREFIX}gcc 
-        export CXX=${RABBITIM_BUILD_CROSS_PREFIX}g++
-        export AR=${RABBITIM_BUILD_CROSS_PREFIX}ar
-        export LD=${RABBITIM_BUILD_CROSS_PREFIX}ld
-        export AS=${RABBITIM_BUILD_CROSS_PREFIX}as
-        export STRIP=${RABBITIM_BUILD_CROSS_PREFIX}strip
-        export NM=${RABBITIM_BUILD_CROSS_PREFIX}nm
-        CONFIG_PARA="CXX=${RABBITIM_BUILD_CROSS_PREFIX}g++ LD=${RABBITIM_BUILD_CROSS_PREFIX}ld"
-        CONFIG_PARA="${CONFIG_PARA} --disable-shared -enable-static --host=$RABBITIM_BUILD_CROSS_HOST"
-        CONFIG_PARA="${CONFIG_PARA} --with-sysroot=${RABBITIM_BUILD_CROSS_SYSROOT}"
-        CFLAGS="-march=armv7-a --sysroot=${RABBITIM_BUILD_CROSS_SYSROOT} "
-        CXXFLAGS="-march=armv7-a --sysroot=${RABBITIM_BUILD_CROSS_SYSROOT} ${RABBITIM_BUILD_CROSS_STL_INCLUDE_FLAGS}"
-        CPPFLAGS=${CXXFLAGS}
-        ;;
+        if [ -n "$RABBITIM_CMAKE_MAKE_PROGRAM" ]; then
+            CMAKE_PARA="${CMAKE_PARA} -DCMAKE_MAKE_PROGRAM=$RABBITIM_CMAKE_MAKE_PROGRAM" 
+        fi
+        CMAKE_PARA="${CMAKE_PARA} -DCMAKE_TOOLCHAIN_FILE=$RABBITIM_BUILD_PREFIX/../../cmake/platforms/toolchain-android.cmake"
+    ;;
     unix)
+        CONFIG_PARA="${CONFIG_PARA} --with-gnu-ld --enable-sse "
         ;;
     windows_msvc)
-        nmake Makefile.vc
-        nmake Makefile.vc install
-        cd $CUR_DIR
-        exit 0
+        MAKE_PARA=""
         ;;
     windows_mingw)
         case `uname -s` in
             Linux*|Unix*|CYGWIN*)
-                CONFIG_PARA="${CONFIG_PARA} CC=${RABBITIM_BUILD_CROSS_PREFIX}gcc --host=${RABBITIM_BUILD_CROSS_HOST} "
-                CONFIG_PARA="${CONFIG_PARA}"
-                ;;
-            MINGW* | MSYS*)
+                CMAKE_PARA="${CMAKE_PARA} -DCMAKE_TOOLCHAIN_FILE=$RABBITIM_BUILD_PREFIX/../../cmake/platforms/toolchain-mingw.cmake"
                 ;;
             *)
             ;;
         esac
         ;;
     *)
-        echo "${HELP_STRING}"
-        cd $CUR_DIR
-        exit 3
-        ;;
+    echo "${HELP_STRING}"
+    cd $CUR_DIR
+    exit 3
+    ;;
 esac
 
-echo "make install"
-echo "pwd:`pwd`"
-CONFIG_PARA="${CONFIG_PARA} --prefix=${RABBITIM_BUILD_PREFIX} "
+echo "cmake .. -DCMAKE_INSTALL_PREFIX=$RABBITIM_BUILD_PREFIX -DCMAKE_BUILD_TYPE=Release -G\"${GENERATORS}\" ${CMAKE_PARA}"
+cmake .. \
+    -DCMAKE_INSTALL_PREFIX="$RABBITIM_BUILD_PREFIX" \
+    -DCMAKE_BUILD_TYPE="Release" \
+    -G"${GENERATORS}" ${CMAKE_PARA} -Dlzma=OFF
 
-echo "./configure ${CONFIG_PARA} CFLAGS=\"${CFLAGS=}\" 
-     CPPFLAGS=\"${CPPFLAGS}\" CXXFLAGS=\"${CXXFLAGS}\" 
-     LDFLAGS=\"$LDFLAGS\" LIBS=\"$LIBS\""
-./configure ${CONFIG_PARA} CFLAGS="${CFLAGS}" \
-    CPPFLAGS="${CPPFLAGS}" CXXFLAGS="${CXXFLAGS}" \
-    LDFLAGS="$LDFLAGS" LIBS="$LIBS"
-
-${MAKE} && ${MAKE} install
-
+cmake --build . --target install --config Release ${MAKE_PARA}
+mkdir -p $RABBITIM_BUILD_PREFIX/lib/pkgconfig
+cp $RABBITIM_BUILD_PREFIX/share/pkgconfig/* $RABBITIM_BUILD_PREFIX/lib/pkgconfig/.
 cd $CUR_DIR
